refactor: refactor packages to controller/model/view
