--- src/main/kotlin/blackjack/Card.kt
@@ -1,7 +1,0 @@
-package blackjack
-
-data class Card(val symbol: Symbol, val index: Int) {
-    init {
-        require(index in 1..13)
-    }
-}
