+++ src/main/kotlin/blackjack/model/Card.kt
@@ -1,0 +1,7 @@
+package blackjack.model
+
+data class Card(val suit: Suit, val index: Int) {
+    init {
+        require(index in 1..13)
+    }
+}
