+++ src/main/kotlin/blackjack/model/Hold.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 class Hold(hold: Set<Card>) {
     constructor() : this(setOf())
