+++ src/main/kotlin/blackjack/model/Person.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 @JvmInline
 value class Person(val name: String)
