+++ src/test/kotlin/blackjack/model/CardTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
@@ -7,7 +7,7 @@ class CardTest {
     @Test
     fun `throw if card is illegal`() {
         assertThrows<IllegalArgumentException> {
-            Card(Symbol.SPADES, 14)
+            Card(Suit.SPADES, 14)
         }
     }
 }
